% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_box.R
\name{bio_box}
\alias{bio_box}
\title{bio_box}
\usage{
bio_box(
  indt,
  pageV = "PARAM",
  visCharV = "VIS",
  visNumV = "VISITNUM",
  xCharV = "TRT01A",
  xNumV = "TRT01AN",
  xColor = c("gray"),
  yV = "AVAL",
  groupV = NULL,
  tit = "Biomarker Box Plot",
  foot3 = "",
  foot4 = "",
  foot5 = "",
  foot6 = "",
  foot7 = "",
  foot8 = "",
  pgName = "bio_box",
  subFolder = NULL,
  groupSize = NULL,
  nSize = NULL,
  visSize = NULL,
  pvalueSize = NULL
)
}
\arguments{
\item{indt}{Required, a R dataset name which is used to create the box plot, such as adrc.}

\item{pageV}{Required, a character string representing the page variable name, such as 'PARAM'.
Each value of page variable will have a box plot on one page.}

\item{visCharV}{Required, a character string representing the character visit variable name, such as 'VISIT'.}

\item{visNumV}{Required, a character string representing the numeric visit variable name, such as 'VISITNUM'.}

\item{xCharV}{Required, a character string representing the character x variable name, such as 'TRT01A'.}

\item{xNumV}{Required, a character string representing the numeric x variable name, such as 'TRT01AN'.}

\item{xColor}{Required, a character vector representing the color for each value of xCharV,
so the length of xColor should be equal to the number of different values of xCharV,
Such as c(rep('purple', 3), 'gray') if there are 4 different values for xCharV.}

\item{yV}{Required, a character string representing the variable name which is used as the y axis variable, such as 'AVAL'.}

\item{groupV}{A character vector representing the group variable names, such as 'STUDYID' or c('STUDYID', 'AGEGR1'),
the default value is NULL.}

\item{tit}{A character string representing the title of output, the default value is 'Biomarker Box Plot'.}

\item{pgName}{A character string representing the program name displayed in the last footnote,
the default value is 'bio_box'}

\item{subFolder}{A character string representing sub folder name where RTF file locates, such as 'Primary', 'CMI', ...
the default value is NULL.}

\item{groupSize}{A floating number representing the font-size of group variable value (groupV) displayed at the top of box plot,
such as 3.5}

\item{nSize}{A floating number representing the font-size of each number displayed at the bottom of box plot, such as 2.5}

\item{visSize}{A floating number representing the font-size of visit variable value (visCharV) displayed at the top of box plot,
such as 2.5}

\item{pvalueSize}{A floating number representing the font-size of the p-values displayed at the top of box plot, such as 1.8}

\item{footn}{A character string representing the nth footnote displayed as footnote, n = 3, ..., 8, the default value is ''.}
}
\description{
This function generates a box plot of biomarker readouts on a logarithmic scale with a base of 10.
At the top of the plot, the p-values of pairwise comparisons between visits are displayed,
and at the bottom, the numbers of each box are shown respectively.
The output is stored in 1 RTF file under REPORT/OUTPUT/Primary under eWISE environment,
or current working path/Primary under local environment.
}
\note{
In the input dataset indt, in addition to the required parameters, there must also be the variable USUBJID.
}
\examples{
\dontrun{

bio_box(
indt = adbm,
pageV = 'PARAM',
visCharV = 'VIS',
visNumV = 'VISITNUM',
xCharV = 'TRT01A',
xNumV = 'TRT01AN',
xColor = c('gray'),
yV = 'AVAL',
)

}

}
