% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_heatmap.R
\name{bio_heatmap}
\alias{bio_heatmap}
\title{bio_heatmap}
\usage{
bio_heatmap(
  indt,
  pageV,
  columnV,
  columnOrdV = NULL,
  columnColor,
  rowV,
  analysisV,
  analysisType = "GM",
  tit = "Biomarker Heatmap",
  foot1 = "",
  foot2 = "",
  foot3 = "",
  foot4 = "",
  foot5 = "",
  foot6 = "",
  foot7 = "",
  foot8 = "",
  pgName = "bio_heatmap",
  subFolder = NULL
)
}
\arguments{
\item{indt}{Required, a R dataset name which is used to create the heatmap, such as adbm.}

\item{pageV}{Required, a character string representing the page variable name, such as 'AGEGR1'.
Each value of page variable will have a heatmap on one page.}

\item{columnV}{Required, a character vector representing the column variable name of heatmap,
such as c('TRT01A', 'VIS') or 'TRT01A'.}

\item{columnOrdV}{A character vector representing the sorting variable name for column,
such as c('TRT01AN', 'VISITNUM') or 'TRT01AN'. The default value is NULL, columnV is used in this case.}

\item{columnColor}{Required, a list representing the colors for column variable name, such as
list(c('orchid3', 'darkorchid1', 'darkorchid2', 'darkorchid3', 'darkorchid4','azure3'), c("#F0F0F0", "#9A9A9A", "#545454")) or
list(c('orchid3', 'darkorchid1', 'darkorchid2', 'darkorchid3', 'darkorchid4','azure3')).
The length of each element of columnColor should be equal to the number of different values of columnV respectively,}

\item{rowV}{Required, a character string representing the row variable name of heatmap, such as 'PARAM'.}

\item{analysisV}{Required, a character string representing the variable name which is used to create heatmap,
such as 'AVAL', 'R2BASE' or 'FOLREISE'.}

\item{analysisType}{A character string representing the type of heapmap.
There are two values available for selection 'GM' or 'Individual', the default value is 'GM'.}

\item{tit}{A character string representing the title of output, the default value is 'Biomarker Heatmap'.}

\item{pgName}{A character string representing the program name displayed in the last footnote,
the default value is 'bio_heatmap'.}

\item{subFolder}{A character string representing sub folder name where RTF file locates, such as 'Primary', 'CMI', ...
the default value is NULL.}

\item{footn}{A character string representing the nth footnote displayed as footnote, n = 1, ..., 8, the default value is ''.}
}
\description{
This function generates a heatmap of biomarker readouts.
The output is stored in 1 RTF file under REPORT/OUTPUT/subFolder under eWISE environment,
or current working path/subFolder under local environment.
}
\note{
In the input dataset indt, in addition to the required parameters, there must also be the variable USUBJID.
}
\examples{
\dontrun{

adbm <- readRDS(system.file("extdata", "example.data.rds", "VacBioTLF")) \%>\%
filter(!is.na(AVAL)) \%>\%
mutate(VIS = str_split_i(VISIT, '/', 2))

bio_heatmap(
indt = adbm,
pageV = 'AGEGR1',
columnV = c('TRT01A', 'VIS'),
columnOrdV = c('TRT01AN', 'VISITNUM'),
columnColor = list(c('orchid3', 'darkorchid1', 'darkorchid2', 'darkorchid3', 'darkorchid4','azure3'),
rowV = 'PARAM',
analysisV = 'AVAL',
analysisType = 'GM'
)

}

}
