% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_spagetti.R
\name{bio_spagetti}
\alias{bio_spagetti}
\title{bio_spagetti}
\usage{
bio_spagetti(
  indt,
  pageV = NULL,
  xV,
  xOrdV = NULL,
  yV,
  columnV = NULL,
  columnOrdV = NULL,
  rowV = NULL,
  rowOrdV = NULL,
  gmColor = NULL,
  tit = "Biomarker Spagetti with GM and 95\% CI",
  foot1 = "",
  foot2 = "",
  foot3 = "",
  foot4 = "",
  foot5 = "",
  foot6 = "",
  foot7 = "",
  foot8 = "",
  pgName = "bio_spagetti",
  subFolder = NULL
)
}
\arguments{
\item{indt}{Required, a R dataset name which is used to create the spagetti, such as adbm.}

\item{pageV}{A character string representing the page variable name, such as 'PARAM'.
Each value of page variable will have a spagetti on one page.
The default value is NULL, which will create only 1 spagetti.}

\item{xV}{Required, a character string representing the X variable name of spagetti, such as 'VISIT'.}

\item{xOrdV}{A character string representing the sorting variable name for X axis, such as 'VISITNUM'.
The default value is NULL, it will be displayed in the order of xV.}

\item{yV}{Required, a character string representing the Y variable name of spagetti, such as 'AVAL'.}

\item{columnV}{a character string representing the column variable name for facet_grid, such as 'TRT01A'.
The default value is NULL.}

\item{columnOrdV}{a character string representing sorting variable name for columnV, such as 'TRT01AN'.
The default value is NULL, it will be displayed in the order of columnV.}

\item{rowV}{a character string representing the column variable name for facet_grid, such as 'AGEGR1'.
The default value is NULL.}

\item{rowOrdV}{a character string representing sorting variable name for rowV, such as 'AGEGR1N'.
The default value is NULL, it will be displayed in the order of rowV}

\item{gmColor}{A character vector representing the color for geometric mean by columnV,
so the number of gmColor should be equal to the number of different value of columnV.
The default value is NULL, it will color randomly.}

\item{tit}{A character string representing the title of output, the default value is 'Biomarker Spagetti'.}

\item{pgName}{A character string representing the program name displayed in the last footnote,
the default value is 'bio_spagetti'.}

\item{subFolder}{A character string representing sub folder name where RTF file locates, such as 'Primary', 'CMI', ...
the default value is NULL.}

\item{footn}{A character string representing the nth footnote displayed as footnote, n = 1, ..., 8,
the default value is ''.}
}
\description{
This function generates a spagetti plot of biomarker readouts.
The output is stored in 1 RTF file under REPORT/OUTPUT/Primary under eWISE environment,
or current working path/Primary under local environment.
}
\note{
In the input dataset indt, in addition to the required parameters, there must also be the variable USUBJID.
}
\examples{
\dontrun{

adbm <- readRDS(here::here("data", "example.data.rds")) \%>\%
filter(!is.na(AVAL)) \%>\%
mutate(VIS = str_split_i(VISIT, '/', 2),
sexn = ifelse(SEX == 'F', 2, 1))

bio_spagetti(
indt = adbm,
pageV = 'PARAM',
xV = 'VIS',
yV = 'AVAL',
columnV = 'TRT01A',
columnOrdV = 'TRT01AN',
rowV = 'SEX',
rowOrdV = 'sexn',
gmColor = c('red', 'yellow')
)

}

}
