% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_spider.R
\name{bio_spider}
\alias{bio_spider}
\title{bio_spider}
\usage{
bio_spider(
  indt,
  pageV = NULL,
  dimV,
  dimOrdV = NULL,
  valueV,
  groupV = NULL,
  groupOrdV = NULL,
  groupColor = c("#FFEC8B", "#C4B045", "#A89222", "#8B7500", "gray35", "gray"),
  refValue = NULL,
  pValueDT = NULL,
  tit = "Spider plot of mean of biomarker readouts",
  foot1 = "",
  foot2 = "",
  foot3 = "",
  foot4 = "",
  foot5 = "",
  foot6 = "",
  foot7 = "",
  foot8 = "",
  pgName = "bio_spider",
  subFolder = NULL
)
}
\arguments{
\item{indt}{Required, a R dataset name which is used to create the spider, such as adbm.}

\item{pageV}{A character string representing the page variable name, such as 'VISIT'.
Each value of page variable will have a spider on one page.
The default value is NULL, which will create only 1 spider.}

\item{dimV}{Required, a character string representing the dimension variable name of spider, such as 'PARAM'.}

\item{dimOrdV}{A character string representing the sorting variable name for dimension variable, such as 'PARAMN'.
The default value is NULL, it will be displayed in the order of dimV.}

\item{valueV}{Required, a character string representing the Value variable name of spider, such as 'AVAL'.}

\item{groupV}{a character string representing the group variable name of spider, such as 'TRT01A'.
The default value is NULL.}

\item{groupOrdV}{a character string representing the sorting order for groupColor.
The default value is NULL.}

\item{groupColor}{a character vector representing the color for group,
so the number of groupColor should be equal to the number of different value of groupV or groupOrdV.
The default value is NULL, it will color randomly.}

\item{refValue}{a floating number representing the reference circle of the spider, the default value is 0.}

\item{pValueDT}{a R dataset name contains the p-values of pairwise comparisons between groupV.
In addition to other necessary variables, variable estimates must be included in the pValueDT.
For example, dimV and groupV, or pageV (if any).}

\item{tit}{A character string representing the title of output, the default value is 'Biomarker Spider'.}

\item{pgName}{A character string representing the program name displayed in the last footnote,
the default value is 'bio_spider'.}

\item{subFolder}{A character string representing sub folder name where RTF file locates, such as 'Primary', 'CMI', ...
the default value is NULL.}

\item{footn}{A character string representing the nth footnote displayed as footnote, n = 1, ..., 8,
the default value is ''.}
}
\description{
This function generates a spider plot of mean of biomarker readouts.
The output is stored in 1 RTF file under REPORT/OUTPUT/subFolder under eWISE environment,
or current working path/subFolder under local environment.
}
\examples{
\dontrun{

adbm <- readRDS(system.file("extdata", "example.data.rds", "VacBioTLF")) \%>\%
filter(!is.na(R2BASE) & VISITNUM > 10) \%>\%
mutate(VIS = str_split_i(VISIT, '/', 2),
sexn = ifelse(SEX == 'F', 2, 1))

pvalue.data <- readRDS(system.file("extdata", "pvalue.data.rds", "VacBioTLF"))

bio_spider(
indt = adbm,
pageV = 'VIS',
dimV = 'PARAM',
dimOrdV = NULL,
valueV = 'R2BASE',
groupV = 'TRT01A',
groupOrdV = 'TRT01AN',
refValue = 1,
pValueDT = pvalue.data
)

}  # Dimension Variable, Value Variable, Grouping Variable

}
