% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imm_summary_gmt.R
\name{imm_summary_gmt}
\alias{imm_summary_gmt}
\title{imm_summary_gmt}
\usage{
imm_summary_gmt(
  indt,
  indt2,
  groupV = "PARAM",
  grpColSize = 6.5,
  timeV = "VIS",
  timeColSize = 3,
  avalV = "AVAL",
  foldV = "FOLDRISE",
  trtnV = "TRT01AN",
  trtcV = "TRT01A",
  pgName = "imm_summary_gmt",
  titV = "",
  foot1V = "",
  foot2V = "",
  foot3V = "",
  foot4V = "",
  foot5V = "",
  foot6V = "",
  foot7V = "",
  foot8V = ""
)
}
\arguments{
\item{indt:}{A R dataset name which contains the immunogenicity data, such as adis}

\item{indt2:}{A R dataset name which used to calculate the big N in the header, such as adsl}

\item{groupV:}{A character vector presents category variable names used in the table and can be found in indt,
such as 'PARAM' or c('AGE', 'PARAM)}

\item{grpColSize:}{A float vector presents the width for category variables, such as 6.5 or c(5, 6.5),
default value is 6.5 for each category variable}

\item{timeV:}{A character string presents visit variable name used in the table and can be found in indt,
such as VISIT}

\item{timeColSize:}{A float number presents the width for visit variable, such as 3}

\item{avalV:}{A character string presents analysis variable name used to calculate GMT and can be found in indt,
the default value is 'AVAL'}

\item{foldV:}{A character string presents fold rise variable name used to calculate GMTR and can be found in indt,
the default value is 'FOLDRISE'}

\item{trtnV:}{A character string presents the numeric treatment name in indt, such as 'TRT01AN'}

\item{trtcV:}{A character string presents the character treatment name in indt and indt2, such as 'TRT01A'.
And there is a 121 relationship between trtnV & trtcV}

\item{pgName:}{A character string presents program name showing in the last footnote in the RTF file,
the default value is 'imm_summary_gmt'}

\item{titV:}{A character string presents the title showing in the RTF file, the default value is ''}

\item{footxV:}{x = 1, ...,8, a character string presents the footnote showing in the RTF file, the default value is ''}
}
\description{
This function generates gmt table for immunogenicity in 1 RTF file
stored under REPORT/OUTPUT/Primary under eWISE environment,
or current working path/Primary under local environment.
}
\note{
There must be variable ATOXGD01, ..., ATOXGD09 in the input dataset indt
}
\examples{
\dontrun{

imm_summary_gmt(indt = final, indt2 = adsl,
groupV = 'PARAM', grpColSize = 6.5,
timeV = 'VIS', timeColSize = 3,
trtnV = 'TRT01AN', trtcV = 'TRT01A',
avalV = 'AVAL', foldV = 'FOLDRISE',
pgName = 'imm_summary_gmt',
titV = 'Table 2.55 Summary of geometric means of antibody titers after vaccination - mFAS',
foot1V = 'M: number of participants with available data for the relevant endpoint.',
foot2V = 'GMT: geometric mean titer. GMTR: geometric mean titer ratio. CI = Confidence interval. NC = Not Computed.',
foot3V = 'The 2 sided 95\% CIs is calculated based on the t-test.',
foot4V = 'For main cohorts, each of the unadjuvanted groups (G10, G11 or G12) are made up of the sum of participants across the main cohorts analyzed.',
)

}

}
