% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacto_bar.R
\name{reacto_bar}
\alias{reacto_bar}
\title{reacto_bar}
\usage{
reacto_bar(
  indt,
  trtnV,
  trtcV,
  tits = c("Figure 1: Barplot for reactogenicity (Max grade over D1-3)",
    "Figure 2: Barplot for reactogenicity (Max grade over D1-9)")
)
}
\arguments{
\item{indt:}{A R dataset name which is used to create the bar plot, such as adrc}

\item{trtnV:}{A character string presents the numeric treatment name in indt, such as 'TRT01AN'}

\item{trtcV:}{A character string presents the character treatment name in indt, such as 'TRT01A'.
And there is a 121 relationship between trtnV & trtcV}

\item{tits:}{titles for the 2 plots, a character vector of length 2.
default value is c('Figure 1: Barplot for reactogenicity (Max grade over D1-3)',
'Figure 2: Barplot for reactogenicity (Max grade over D1-9)')}
}
\description{
This function generates 2 reactogenicity barplots for Max grade over D1-3 and
Max grade over D1-9 in 1 RTF file stored under REPORT/OUTPUT/Primary under eWISE environment,
or current working path/Primary under local environment.
}
\note{
There must be variable ATOXGD01, ..., ATOXGD09 in the input dataset indt
}
\examples{
\dontrun{

reacto_bar(indt = adrc,
trtnV ='TRT01AN',
trtcV = 'TRT01AC',
tits = c('Figure 1: Barplot for reactogenicity (Max grade over D1-3)',
'Figure 2: Barplot for reactogenicity (Max grade over D1-9)'))

}

}
