% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_example_data.R
\docType{data}
\name{subCnt}
\alias{subCnt}
\title{example_data}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
subCnt
}
\description{
This function generates an example data for function to use
}
\keyword{datasets}
